requirements.txt
pillow
scipy